cap nhat file ban , chitiethoadon , hoadon bus dap dto

+++ DTO/SanPhamDTO.cs
@@ -10,14 +10,14 @@
         {
             public int Id_SanPham { get; set; }
             public string TenSanPham { get; set; }
-            public decimal GiaMua { get; set; }
+            public float GiaMua { get; set; }
             public int SoLuongTon { get; set; }
             public string TrangThai { get; set; }
             public int Id_DanhMuc { get; set; }
 
             public SanPhamDTO() { }
 
-            public SanPhamDTO(int id_SanPham, string tenSanPham, decimal giaMua, int soLuongTon, string trangThai, int id_DanhMuc)
+            public SanPhamDTO(int id_SanPham, string tenSanPham, float giaMua, int soLuongTon, string trangThai, int id_DanhMuc)
             {
                 Id_SanPham = id_SanPham;
                 TenSanPham = tenSanPham;
